divide into elements
