+++ client/src/components/message/message.scss
@@ -1,0 +1,11 @@
+.message-style {
+  display: block;
+  border-style: solid;
+  border-radius: 25px;
+  text-align: center;
+  border-color: #cd3c45;
+  font-family: 'Lucida Console', 'Courier New', monospace;
+  color: #cd3c45;
+  font-size: 10px;
+  margin-bottom: 50px;
+}
