+++ client/src/consts.js
@@ -1,0 +1,3 @@
+import React from 'react';
+window.baseUrl = 'http://localhost:3000/'
+window.singInUrl = window.baseUrl + 'users/sign_in'
