+++ client/src/index.js
@@ -1,19 +1,9 @@
 import ReactDOM from "react-dom";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Layout from "./pages/Layout";
-import SingIn from "./pages/SingIn";
-import NoPage from "./pages/NoPage";
+import CustomRoutes from "./components/custom-routes/custom-routes";
 
 export default function App() {
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Layout />}>
-          <Route path="/sign_in" element={<SingIn />} />
-          <Route path="*" element={<NoPage />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
+    <CustomRoutes/>
   );
 }
