--- client/src/pages/NoPage.js
@@ -1,13 +1,0 @@
-const style = {
-  marginTop: "150px",
-  textAlign: "center",
-  fontFamily: "'Lucida Console', 'Courier New', monospace",
-  fontSize: "90px",
-  color: "dodgerblue",
-};
-
-const NoPage = () => {
-  return <h1 style={style}>404</h1>;
-};
-
-export default NoPage;
