+++ client/src/pages/NoPage/NoPage.js
@@ -1,0 +1,11 @@
+import './NoPage.scss'
+
+const NoPage = () => {
+  return (
+    <div className='error-code'>
+      404
+    </div>
+  );
+};
+
+export default NoPage;
