+++ client/src/pages/NoPage/NoPage.scss
@@ -1,0 +1,7 @@
+.error-code {
+  margin-top: 150px;
+  text-align: center;
+  font-family: 'Lucida Console', 'Courier New', monospace;
+  font-size: 90px;
+  color: dodgerblue;
+}
